Include all headers regardless of depth
The previous hack for this was to move headers 3 levels deep in to
the top level include dir and headers 4 levels deep in to sys.
Obviously this was fairly limiting and missed two important headers.

  ./zfs/include/sys/fm/fs/zfs.h
  ./zfs/include/sys/fs/zfs.h

This patch updates the code to be smarter and installs all the headers
to the correct location regardless of depth.  Each header will have its
leading 3 path args stripped (i.e. ./zfs/include/) and replaced with
the correct install destination.  All path information past the first
three levels will be preserved.