Fix struct ht_lock padding in arc.c.
This was leading to a compilation error because in Linux, sizeof
(kmutex_t) can be larger than 64 in some circumstances (e.g. with
debugging and lockdep enabled).

The code was previously fixed to align the structure to 256 bytes,
but a better fix is to align it to the next multiple of 64 bytes.