Revert zfs_ioctl.c related changes.
All the changes required to cleanly integrate with linux style
devices are being moved to their own topic branch.  It was abouti
to get to the point where too many topic branches each wanted to
tweak this code and it was getting confusing.  To simplify this
all /dev/zfs device related changes are being moved to the
linux-kernel-device topic branch.