Remove explicit import from zconfig test 1.
Due to now resolved bug in the SPL you would need to explicitly
import you zpools after module load.  Now that is no longer the case.
If a cache file is found your pool will be automatically loaded and
available so I'm removing the explicit imports from the test case.