Fix error message to reference the new libuuid package.
It used to be the case that libuuid was part of e2fsprogs.  However,
since other projects unrelated to e2fsprogs make use of that library
it was getting aukward to have that dependency.  The package was
split from e2fsprogs and moved to its own devel package in the
util-linux-ng project.  Anyway, I'm just updating the error message
to reflect the new upstream reality.