Move from kmem_alloc() to vmem_alloc().
These two allocations are 16k in size which trips the SPL warning
about large kmem_allocs().  For now simply shift them to a vmem_alloc().
Yes, this has it's own set of problems but this happens infrequently
enough not to be a real issue.  But more importantly it means that
we've flagged this place in the code via this topic branch as a
place which still needs long term work.