Prep for 0.4.9 tag update META, README, ChangeLog.
Updated README to reflect the current state of the project and to
be less developer focused.  It now simply includes basic build
instructions for packages and little example of some of the cool
things you can do.

Also remove the out of data TODO file.  This sort of thing can
and be tracked on the GitHub issue tracker.