Silence 3 additional large kmem warnings
The following are 3 cases where move than 2 pages are allocated
with a kmem_alloc()... but not a lot more.  For now we just disable
the warning with KM_NODEBUG and this can be revisted latter to
see if it's worth shrinking the allocation or perhaps moving it
to a slab.